Controllo accessi user, aggiunta comunicazioni

+++ src/main/java/com/marco/unicorsi/controllers/UserController.java
@@ -2,10 +2,12 @@ package com.marco.unicorsi.controllers;
 
 import java.security.Principal;
 
+import com.marco.unicorsi.model.Comunicazione;
 import com.marco.unicorsi.model.Corso;
 import com.marco.unicorsi.model.Lezione;
 import com.marco.unicorsi.model.Professore;
 import com.marco.unicorsi.model.User;
+import com.marco.unicorsi.repository.ComunicationRepo;
 import com.marco.unicorsi.repository.CorsoRepo;
 import com.marco.unicorsi.repository.LessonRepo;
 import com.marco.unicorsi.repository.UserRepo;
@@ -13,6 +15,7 @@ import com.marco.unicorsi.repository.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -35,6 +38,9 @@ public class UserController{
     @Autowired
     LessonRepo lessonRepo;
 
+    @Autowired
+    ComunicationRepo comRepo;
+
     @GetMapping(value={"/index","/"} )
     public ModelAndView getUsers(Principal principal) {
         //L'oggetto va instanziato nei metodi, infatti di ogni classe vi è una singola istanza
@@ -52,17 +58,17 @@ public class UserController{
         model.addObject("isLogged", false);
         model.setViewName("index");
         return model;
-    }
+    }//TODO rivedere metodo
 
     @GetMapping(value="/user-home")
-    public ModelAndView getUserHome(Principal principal) {
-        ModelAndView mView = new ModelAndView();
-        mView.setViewName("/user/user-home");
+    public String getUserHome(Principal principal, Model model) {
         Professore pr = userRepo.findByUsername(principal.getName()).getDocente();
         if(pr!=null) {
-            mView.addObject("docente", pr);
+            model.addAttribute("docente", pr);
+        } else {
+            return "redirect: /admin/admin-home";
         }
-        return mView;
+        return "/user/user-home";
     }
 
     @GetMapping(value="/update-corso")
@@ -70,10 +76,11 @@ public class UserController{
         ModelAndView mView = new ModelAndView();
         //La modifica del corso è consentita solo all'utente titolare del corso
         Corso corso = corsoRepo.findByInsegnamentoAndAnnoAccademico(codice, anno);
-        User user = userRepo.findByUsername(principal.getName());
-        if(corso.getTitolari().contains(user.getDocente())){
+        //User user = userRepo.findByUsername(principal.getName());
+        if(isOwner(principal, corso)){
             mView.setViewName("/user/update-corso");
             mView.addObject("newLez", new Lezione(corso));
+            mView.addObject("newCom", new Comunicazione(corso));
             mView.addObject("corso", corso);
         } else {
             mView.setViewName("/error");
@@ -82,27 +89,41 @@ public class UserController{
     }
 
     @PostMapping(value = "/add-lezione")
-    public String postAddLezione(@ModelAttribute Lezione lezione, RedirectAttributes redirectAttributes){
-        lessonRepo.save(lezione);
-        redirectAttributes.addAttribute("codice", lezione.getCorso().getInsegnamento().getCodice());
-        redirectAttributes.addAttribute("anno", lezione.getCorso().getAnnoAccademico());
-        return "redirect:/user/update-corso";
+    public String postAddLezione(@ModelAttribute Lezione lezione, RedirectAttributes redirectAttributes, Principal principal){
+        Corso corso = lezione.getCorso();
+        if(isOwner(principal, corso)){
+            lessonRepo.save(lezione);
+            redirectAttributes.addAttribute("codice", corso.getInsegnamento().getCodice());
+            redirectAttributes.addAttribute("anno", corso.getAnnoAccademico());
+            return "redirect:/user/update-corso";
+        } else {
+            return "/error";
+        }
     }
 
     @PostMapping(value = "/delete-lesson")
-    public String deleteLezione(@RequestParam String codice, @RequestParam String anno, @RequestParam int idLezione, RedirectAttributes redirectAttributes){
-        redirectAttributes.addAttribute("codice", codice);
-        redirectAttributes.addAttribute("anno", anno);
-        lessonRepo.deleteById(idLezione);
-        return "redirect:/user/update-corso";
+    public String deleteLezione(@RequestParam String codice, @RequestParam String anno, 
+         @RequestParam int idLezione, RedirectAttributes redirectAttributes, Principal principal){
+        Corso corso = corsoRepo.findByInsegnamentoAndAnnoAccademico(codice, anno);
+        if(isOwner(principal, corso)){
+            redirectAttributes.addAttribute("codice", codice);
+            redirectAttributes.addAttribute("anno", anno);
+            lessonRepo.deleteById(idLezione);
+            return "redirect:/user/update-corso";
+        } else {
+            return "/error";
+        }
     }
 
     @PostMapping(value = "/update-prog")
-    public String updateProg(@RequestParam String programma, @RequestParam int idCorso, RedirectAttributes rAttributes){
+    public String updateProg(@RequestParam String programma, @RequestParam int idCorso,
+            RedirectAttributes rAttributes, Principal principal){
         try {
             //L'oggetto corso non viene passato nella chiamata, occorre ritrovarlo nel db per salvarlo
             Corso corso = corsoRepo.findById(idCorso).get();
             corso.setProgramma(programma);
+            if(!isOwner(principal, corso))
+                return "/error";
             rAttributes.addAttribute("codice", corso.getInsegnamento().getCodice());
             rAttributes.addAttribute("anno", corso.getAnnoAccademico());
             corsoRepo.save(corso);
@@ -111,6 +132,28 @@ public class UserController{
         }
         return "redirect:/user/update-corso";
     }
+
+    @PostMapping(value = "/add-com")
+    public String postAddLezione(@ModelAttribute Comunicazione comunicazione, RedirectAttributes redirectAttributes, Principal principal){
+        Corso corso = comunicazione.getCorso();
+        if(isOwner(principal, corso)){
+            comRepo.save(comunicazione);
+            redirectAttributes.addAttribute("codice", corso.getInsegnamento().getCodice());
+            redirectAttributes.addAttribute("anno", corso.getCorso().getAnnoAccademico());
+            return "redirect:/user/update-corso";
+        } else {
+            return "/error";
+        }
+    }
+
+    //Controlla se chi sta eseguendo la richiesta è titolare del corso
+    private boolean isOwner(Principal principal, Corso corso){
+        User user = userRepo.findByUsername(principal.getName());
+        if(corso.getTitolari().contains(user.getDocente()))
+            return true;
+        return false;
+    }
+   
     
 
 }
